+++ SC/View/ViewController/CalculateViewController.swift
@@ -20,7 +20,12 @@ class CalculateViewController: UIViewController {
     @IBOutlet private weak var resultLabel: UILabel!
     
     private var numOnScreen: Float = 0
+    private var preNum: Float = 0
+    
     private var canCalculate: Bool = false
+    private var operationNum: Int = 0
+    
+    private var memoryNumOnScreen: Float = 0
     
     private let greenButtonColor = UIColor(red: 0.8, green: 1.0, blue: 0.8, alpha: 1.0)
     private let url = NSURL(string: PSCStringStorage.init().BLOG_URL)
@@ -51,6 +56,7 @@ class CalculateViewController: UIViewController {
         }
     }
     
+    // Green Buttons: tag = (0~9)
     @IBAction private func numberButtonTapped(_ sender: UIButton) {
         if self.canCalculate == true || self.resultLabel.text == "0" {
             self.resultLabel.text = String(sender.tag)
@@ -60,6 +66,205 @@ class CalculateViewController: UIViewController {
             self.resultLabel.text = self.resultLabel.text! + String(sender.tag)
             self.numOnScreen = NSString(string: self.resultLabel.text!).floatValue
         }
+    }
+    
+    // Dark Gray Buttons: tag = (10~14)
+    // C, ., AC, +-, %
+    @IBAction private func darkGrayOperationButtonTapped(_ sender: UIButton) {
+        switch sender.tag {
+        case 10: // C
+            if self.resultLabel.text!.count > 1 {
+                _ = self.resultLabel.text!.popLast()
+                self.numOnScreen = NSString(string: self.resultLabel.text!).floatValue
+            } else {
+                self.resultLabel.text = "0"
+                self.numOnScreen = 0
+            }
+        case 11: // .
+            if self.getLastChar() != "." {
+                self.resultLabel.text?.append(".")
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+        case 12: // AC
+            self.resultLabel.text = "0"
+            self.preNum = 0
+            self.numOnScreen = 0
+            self.operationNum = 0
+        case 13: // +/-
+            self.preNum = NSString(string: self.resultLabel.text!).floatValue
+            var tmp = NSString(string: self.resultLabel.text!).floatValue
+            tmp *= -1.0
+            self.numOnScreen = tmp
+            self.resultLabel.text = String(tmp)
+        case 14: // %
+            self.calculateOperation() // 最後がoperationじゃないかの確認
+            if self.getLastChar() >= "0" && self.getLastChar() <= "9" {
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+            self.preNum = NSString(string: self.resultLabel.text!).floatValue
+            self.resultLabel.text?.append("%")
+            self.operationNum = sender.tag
+            self.canCalculate = true
+        default:
+            print("may not be come in here...")
+        }
+    }
+    
+    // Yellow Buttons: tag = (15~19)
+    // =, +, -, *, /
+    @IBAction private func yellowOperationButtonTapped(_ sender: UIButton) {
+        switch sender.tag {
+        case 15: // =
+            self.calculateOperation()
+        case 16: // +
+            self.calculateOperation() // 最後がoperationじゃないかの確認
+            if self.getLastChar() >= "0" && self.getLastChar() <= "9" {
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+            self.preNum = NSString(string: self.resultLabel.text!).floatValue
+            self.resultLabel.text?.append("+")
+            self.operationNum = sender.tag
+            self.canCalculate = true
+        case 17: // -
+            self.calculateOperation() // 最後がoperationじゃないかの確認
+            if self.getLastChar() >= "0" && self.getLastChar() <= "9" {
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+            self.preNum = NSString(string: self.resultLabel.text!).floatValue
+            self.resultLabel.text?.append("-")
+            self.operationNum = sender.tag
+            self.canCalculate = true
+        case 18: // ×
+            self.calculateOperation() // 最後がoperationじゃないかの確認
+            if self.getLastChar() >= "0" && self.getLastChar() <= "9" {
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+            self.preNum = NSString(string: self.resultLabel.text!).floatValue
+            self.resultLabel.text?.append("×")
+            self.operationNum = sender.tag
+            self.canCalculate = true
+        case 19: // ÷
+            self.calculateOperation() // 最後がoperationじゃないかの確認
+            if self.getLastChar() >= "0" && self.getLastChar() <= "9" {
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+            self.preNum = NSString(string: self.resultLabel.text!).floatValue
+            self.resultLabel.text?.append("÷")
+            self.operationNum = sender.tag
+            self.canCalculate = true
+        default:
+            print("may not be come in here...")
+        }
+    }
+    
+    // Light Gray Buttons: tag = (20~29)
+    // √, !, 1/x, ^x, 10^x, MC, M+, M-, MR, ?
+    @IBAction private func lightGrayOperationButtonTapped(_ sender: UIButton) {
+        switch sender.tag {
+        case 20: // √
+            // TODO: perationじゃないかの確認 する？
+            if self.getLastChar() >= "0" && self.getLastChar() <= "9" {
+                self.preNum = NSString(string: self.resultLabel.text!).floatValue
+                self.resultLabel.text = String(sqrtf(self.numOnScreen))
+                self.numOnScreen = NSString(string: self.resultLabel.text!).floatValue
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+        case 21: // !
+            // TODO: perationじゃないかの確認 する？
+            if self.resultLabel.text == "0" || self.resultLabel.text == "0.0" {
+            } else {
+                self.preNum = NSString(string: self.resultLabel.text!).floatValue
+                let roop = Int(self.numOnScreen)
+                var ans = 1
+                for i in 0..<roop {
+                    ans *= i+1
+                }
+                self.resultLabel.text = String(ans)
+                self.numOnScreen = NSString(string: self.resultLabel.text!).floatValue
+            }
+        case 22: // 1/x
+            // TODO: perationじゃないかの確認 する？
+            if self.resultLabel.text == "0" || self.resultLabel.text == "0.0" {
+            } else {
+                self.preNum = NSString(string: self.resultLabel.text!).floatValue
+                self.resultLabel.text = String(1/self.numOnScreen)
+                self.numOnScreen = NSString(string: self.resultLabel.text!).floatValue
+            }
+        case 23: // ^x
+            self.calculateOperation() // 最後がoperationじゃないかの確認
+            if self.getLastChar() >= "0" && self.getLastChar() <= "9" {
+            } else {
+                _ = self.resultLabel.text?.popLast()
+            }
+            self.preNum = NSString(string: self.resultLabel.text!).floatValue
+            self.resultLabel.text?.append("^")
+            self.operationNum = sender.tag
+            self.canCalculate = true
+        case 24: // 10^x
+            // TODO: perationじゃないかの確認 する？
+            if self.resultLabel.text == "0" || self.resultLabel.text == "0.0" {
+            } else {
+                self.preNum = NSString(string: self.resultLabel.text!).floatValue
+                self.resultLabel.text = String(powf(10, self.numOnScreen))
+                self.numOnScreen = NSString(string: self.resultLabel.text!).floatValue
+            }
+        case 25: // mc
+            self.memoryNumOnScreen = 0
+            self.memoryLabel.text = String(self.memoryNumOnScreen)
+        case 26: // m+
+            self.memoryNumOnScreen += self.numOnScreen
+            self.memoryLabel.text = String(self.memoryNumOnScreen)
+        case 27: // m-
+            self.memoryNumOnScreen -= self.numOnScreen
+            self.memoryLabel.text = String(self.memoryNumOnScreen)
+        case 28: // mr
+            self.numOnScreen = self.memoryNumOnScreen
+            self.resultLabel.text = String(self.numOnScreen)
+        case 29: // ?
+            if UIApplication.shared.canOpenURL(self.url! as URL) {
+                UIApplication.shared.open(self.url! as URL, options: [:], completionHandler: nil)
+            }
+        default:
+            print("may not be come in here...")
+        }
+    }
+    
+    private func calculateOperation() {
+        switch self.operationNum {
+        case 14: // %
+            self.resultLabel.text = String(self.preNum.truncatingRemainder(dividingBy: self.numOnScreen))
+            self.numOnScreen = self.preNum.truncatingRemainder(dividingBy: self.numOnScreen)
+        case 16: // +
+            self.resultLabel.text = String(self.preNum + self.numOnScreen)
+            self.numOnScreen += self.preNum
+        case 17: // -
+            self.resultLabel.text = String(self.preNum - self.numOnScreen)
+            self.numOnScreen -= self.preNum
+        case 18: // *
+            self.resultLabel.text = String(self.preNum * self.numOnScreen)
+            self.numOnScreen *= self.preNum
+        case 19: // /
+            self.resultLabel.text = String(self.preNum / self.numOnScreen)
+            self.numOnScreen /= self.preNum
+        case 23: // ^x
+            self.resultLabel.text = String(powf(self.preNum, self.numOnScreen))
+            self.numOnScreen = powf(self.preNum, self.numOnScreen)
+        default:
+            print("may not be come in here.")
+        }
+        
+        self.operationNum = 0
+    }
+    
+    private func getLastChar() -> Character {
+        return resultLabel.text?.last ?? "0"
     }
     
 }
